Refactor jinja templatiziation to common location.
